+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Development
 - Support MONGODB-AWS authentication mechanism (with `authmechanismproperties`) #2507
 - Turning off dereferencing for the results of distinct query. #2663
 - Add tests against Mongo 5.0 in pipeline
+- Drop support for Python 3.6 (EOL)
 
 Changes in 0.24.2
 =================